--- .env.development
@@ -1,1 +1,0 @@
-BLOG_URL=/
