--- .env.production
@@ -1,1 +1,0 @@
-BLOG_URL=https://blog.smockle.com/
