+++ lambda/smockle-origin-request.js
@@ -2,6 +2,10 @@ function removeBlog(x) {
   return x.replace(/(\/\/[\w.]+)\/blog/, "$1");
 }
 
+function addBlog(x) {
+  return x.replace(/(\/\/[\w.]+)\//, "$1/blog/");
+}
+
 function addTrailingSlash(x) {
   const [, match] = /((.*[^\/]$)|(.*(?=[\/]$)))/.exec(x);
   return `${match}/`;
@@ -9,6 +13,8 @@ function addTrailingSlash(x) {
 
 exports.handler = async event => {
   var request = event.Records[0].cf.request;
-  request.uri = addTrailingSlash(removeBlog(request.uri));
+  if (request.uri.endsWith("/")) {
+    request.uri = addTrailingSlash(addBlog(removeBlog(request.uri)));
+  }
   return request;
 };
