+++ src/components/header.tsx
@@ -1,10 +1,11 @@
 import React from "react";
+import { Link } from "@reach/router";
 
 export function Header() {
   return (
     <header role="banner">
       <div>
-        <a rel="author" href={process.env.BLOG_URL}>
+        <Link rel="author" to="/blog/">
           <img
             className="logo"
             alt="logo"
@@ -12,7 +13,7 @@ export function Header() {
             height="48"
             width="48"
           />
-        </a>
+        </Link>
       </div>
     </header>
   );
